+++ unicorn_binance_websocket_api/connection.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
+# ¯\_(ツ)_/¯
 #
 # File: unicorn_binance_websocket_api/connection.py
 #
@@ -20,12 +21,14 @@
 
 from .exceptions import *
 from urllib.parse import urlparse
+
 import asyncio
 import copy
 import logging
 import socks  # PySocks https://pypi.org/project/PySocks/
 import sys
 import websockets
+
 
 __logger__: logging.getLogger = logging.getLogger("unicorn_binance_websocket_api")